creates initial nginx+supervisor config
To support legacy redirect patterns we need to run a web server in
front of Fastboot. The easiest way to do this is to run an Nginx
instance in the same container as Fastboot (using supervisord to manage
the processes). This commit provides the initial nginx + supervisor
config and updates the Dockerfile accordingly.

NOTE: This is currently sans-redirects.