uses root user for fastboot service (temporary)
This commit uses the root user for the fastboot service.
Generally, this should be fine because we aren't using volume mounts,
etc... but best practice probably says that we change this at some
point.

The issue is that the `node` user doesn't have access to wherever it's
trying to cache assets from S3.