refactor how the legacy content is parsed
trying to address a few issues here.

it's nice that we can pass `DocumentFragment` directly into a template 
and ember will handle the rest.

of course, under the hood, it's using DOM APIs like `appendChild` or 
`insertBefore` (haven't actually read the source) to move the nodes into 
the document. one of the side effects of these methods is that the 
source `DocumentFragment` is left empty. The result is that when we 
generate a `DocumentFragment` as computed property, it can only render 
out once. Subsequent accesses to that property will return the empty 
`DocumentFragment`.

To get around this, I am accessing the `DocumentFragment` through a 
public getter, which uses `cloneNode` to return a fresh copy of the 
`DocumentFragment` that can be safely emptied into the live document.

There's some additional processing that's needed for MT weirdness. I 
packaged that up inside the the `_parsedLegacyContent` computed 
property. I broke up the lead image and credit/caption in order to keep 
things more functional within the computed property. I'd lik to keep 
things functional from the perspective of the computed, so that any work 
required should return new or update values that can be assembled into a 
parsed object as needed (to whatever extent possible).