tweak how query-more should be initialized
we're getting *more* of something, so the expectation should be that we 
pass in where we start, which would be the second page. the assumption 
is that there's a set that's beging from elsewhere.

regardless, the idea is that this thing will make the query with the 
given `page` value and increment it after retrieving.