move this into the nypr-o-article-body component
it's a little weird, because on the one hand, I'd want the article body
component to be kind of dumb, and just provide the markup and classes so things
look good, so there shouldn't be any reliance on third-party scripts
baked in

on the the other hand, embedded content is very much part of an article body,
so makes a certain amount of sense to wrap it all up.

the timing on didTransition wasn't late enough to call load for
twitter embeds to properly render. moving it to the `didRender` hook of
the component provides a better timing