we can use a document fragment for this
the main win for imgix is on the hoepage. by setting the alt and 
aria-label attributes to values that don't require parsing the body, we 
get around having the browser kick off requests for any images found in 
article bodies when it's run through dom-fixer. this delays running dom 
fixer until an article is actually visited by a user